refactor(components): 폴더 구조 및 css 파일 분리
