+++ src/components/RestaurantFilter/restaurant-filter.css
@@ -1,0 +1,22 @@
+.restaurant-filter-container {
+  display: flex;
+  justify-content: space-between;
+
+  padding: 0 16px;
+  margin-top: 24px;
+}
+
+.restaurant-filter-container select {
+  height: 44px;
+  min-width: 125px;
+
+  border: 1px solid var(--grey-200);
+  border-radius: 8px;
+  background: transparent;
+
+  font-size: 1rem;
+}
+
+.restaurant-filter {
+  padding: 8px;
+}
