+++ src/styles/index.css
@@ -1,3 +1,8 @@
+@import url('../components/Header/header.css');
+@import url('../components/AddRestaurantModal/add-restaurant-modal.css');
+@import url('../components/RestaurantFilter/restaurant-filter.css');
+@import url('../components/Restaurant/restaurant.css');
+
 * {
   padding: 0;
   margin: 0;
@@ -49,255 +54,4 @@ body {
   font-size: 0.875rem;
   line-height: 1.25rem;
   font-weight: 400;
-}
-
-/* GNB ********************************************/
-.gnb {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  height: 64px;
-
-  padding: 0 16px;
-
-  background-color: var(--primary-color);
-}
-
-.gnb__title {
-  color: #fcfcfd;
-}
-
-.gnb__button {
-  height: 40px;
-
-  border: none;
-  border-radius: 8px;
-  background: transparent;
-
-  font-size: 1.5rem;
-  cursor: pointer;
-}
-
-.gnb__button img {
-  display: block;
-  width: 40px;
-  height: 40px;
-  object-fit: contain;
-}
-
-/* 음식점 목록 *****************************************/
-
-/* 카테고리/정렬 필터 */
-.restaurant-filter-container {
-  display: flex;
-  justify-content: space-between;
-
-  padding: 0 16px;
-  margin-top: 24px;
-}
-
-.restaurant-filter-container select {
-  height: 44px;
-  min-width: 125px;
-
-  border: 1px solid #d0d5dd;
-  border-radius: 8px;
-  background: transparent;
-
-  font-size: 1rem;
-}
-
-.restaurant-filter {
-  padding: 8px;
-}
-
-/* 음식점 목록 */
-.restaurant-list-container {
-  display: flex;
-  flex-direction: column;
-
-  padding: 0 16px;
-  margin: 16px 0;
-}
-
-.restaurant {
-  display: flex;
-  align-items: flex-start;
-
-  padding: 16px 8px;
-
-  border-bottom: 1px solid #e9eaed;
-}
-
-.restaurant__category {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 64px;
-  height: 64px;
-  min-width: 64px;
-  min-height: 64px;
-
-  margin-right: 16px;
-
-  border-radius: 50%;
-  background: var(--lighten-color);
-}
-
-.category-icon {
-  width: 36px;
-  height: 36px;
-}
-
-.restaurant__info {
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-}
-
-.restaurant__name {
-  margin: 0;
-}
-
-.restaurant__distance {
-  color: var(--primary-color);
-}
-
-.restaurant__description {
-  display: -webkit-box;
-
-  padding-top: 8px;
-
-  overflow: hidden;
-  text-overflow: ellipsis;
-  -webkit-line-clamp: 2;
-  -webkit-box-orient: vertical;
-}
-
-/* 음식점 추가 모달 *****************************************/
-.modal {
-  display: none;
-}
-
-.modal--open {
-  display: block;
-}
-
-.modal-backdrop {
-  position: fixed;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-
-  background: rgba(0, 0, 0, 0.35);
-}
-
-.modal-container {
-  position: fixed;
-  bottom: 0;
-  width: 100%;
-
-  padding: 32px 16px;
-
-  border-radius: 8px 8px 0px 0px;
-  background: var(--grey-100);
-}
-
-.modal-title {
-  margin-bottom: 36px;
-}
-
-.error-message {
-  height: 20px;
-  color: tomato;
-  font-weight: 600;
-}
-
-.form-item {
-  display: flex;
-  flex-direction: column;
-  margin-bottom: 36px;
-}
-
-.form-item label {
-  color: var(--grey-400);
-  font-size: 0.875rem;
-}
-
-.form-item--required label::after {
-  padding-left: 4px;
-
-  color: var(--primary-color);
-  content: '*';
-}
-
-.form-item .help-text {
-  color: var(--grey-300);
-}
-
-.form-item input,
-.form-item textarea,
-.form-item select {
-  padding: 8px;
-  margin: 6px 0;
-
-  border: 1px solid var(--grey-200);
-  border-radius: 8px;
-
-  font-size: 1rem;
-}
-
-.form-item textarea {
-  resize: none;
-}
-
-.form-item select {
-  height: 44px;
-
-  padding: 8px;
-
-  border: 1px solid var(--grey-200);
-  border-radius: 8px;
-
-  color: var(--grey-300);
-}
-
-input[name='name'],
-input[name='link'] {
-  height: 44px;
-}
-
-.button-container {
-  display: flex;
-}
-
-.button {
-  width: 100%;
-  height: 44px;
-
-  margin-right: 16px;
-
-  border: none;
-  border-radius: 8px;
-
-  font-weight: 600;
-  cursor: pointer;
-}
-
-.button:last-child {
-  margin-right: 0;
-}
-
-.button--secondary {
-  border: 1px solid var(--grey-300);
-  background: transparent;
-
-  color: var(--grey-300);
-}
-
-.button--primary {
-  background: var(--primary-color);
-
-  color: var(--grey-100);
 }
